requirements.txt
torch>=2.2
hydra-core>=1.3
omegaconf>=2.3
wandb>=0.16
matplotlib
joblib
pytest